Created OMERO Extension Menu
