+++ qupath-extension-omero/src/main/java/qupath/lib/images/servers/omero/OmeroWebImageServerBrowserCommand.java
@@ -41,6 +41,7 @@ import javafx.embed.swing.SwingFXUtils;
 import javafx.geometry.Insets;
 import javafx.geometry.Pos;
 import javafx.scene.Node;
+import javafx.scene.Scene;
 import javafx.scene.canvas.Canvas;
 import javafx.scene.canvas.GraphicsContext;
 import javafx.scene.control.Button;
@@ -71,6 +72,7 @@ import javafx.scene.layout.AnchorPane;
 import javafx.scene.layout.BorderPane;
 import javafx.scene.layout.GridPane;
 import javafx.scene.layout.Priority;
+import javafx.stage.Stage;
 import javafx.util.StringConverter;
 import qupath.lib.common.ThreadTools;
 import qupath.lib.gui.QuPathGUI;
@@ -79,6 +81,7 @@ import qupath.lib.gui.dialogs.Dialogs;
 import qupath.lib.gui.tools.GuiTools;
 import qupath.lib.gui.tools.IconFactory;
 import qupath.lib.gui.tools.PaneTools;
+import qupath.lib.images.servers.ImageServer;
 import qupath.lib.images.servers.omero.OmeroObjects.Dataset;
 import qupath.lib.images.servers.omero.OmeroObjects.Group;
 import qupath.lib.images.servers.omero.OmeroObjects.Image;
@@ -87,9 +90,9 @@ import qupath.lib.images.servers.omero.OmeroObjects.Owner;
 import qupath.lib.images.servers.omero.OmeroObjects.Project;
 import qupath.lib.images.servers.omero.OmeroObjects.Server;
 
-class OmeroWebImageServerBrowser {
+class OmeroWebImageServerBrowserCommand implements Runnable {
 	
-	final private static Logger logger = LoggerFactory.getLogger(OmeroWebImageServerBrowser.class);
+	final private static Logger logger = LoggerFactory.getLogger(OmeroWebImageServerBrowserCommand.class);
 	
 	private QuPathGUI qupath;
 	private OmeroWebImageServer server;
@@ -106,7 +109,6 @@ class OmeroWebImageServerBrowser {
 	private int imgPrefSize = 256;
 	private ProgressIndicator progressIndicator;
 	private Button openBtn;
-	private Button refreshBtn;
 	
 	private StringConverter<Owner> ownerStringConverter;
 	
@@ -129,14 +131,38 @@ class OmeroWebImageServerBrowser {
 	private Integer[] projectIndices;
 	
 	
-    OmeroWebImageServerBrowser(QuPathGUI qupath, OmeroWebImageServer server) {
+    public OmeroWebImageServerBrowserCommand(QuPathGUI qupath) {
     	this.qupath = qupath;
-    	this.server = server;
-
+    }
+    
+    @Override
+    public void run() {
+    	// Need to choose which server to browse first
+    	if (qupath.getImageData() != null) {
+    		var serverTemp = qupath.getImageData().getServer();
+    		if (serverTemp instanceof OmeroWebImageServer)
+    			server = (OmeroWebImageServer) serverTemp;
+    	}
+    	while (server == null) {
+//    		String[] choices = qupath.getProject().getImageList().parallelStream().map(e -> e.getImageName()).toArray(String[]::new);
+    		var entry = Dialogs.showChoiceDialog("No open OMERO image", "Choose an entry for a server to browse", qupath.getProject().getImageList(), null);
+    		if (entry == null)
+    			return;
+			try {
+				ImageServer<BufferedImage> serverTemp = entry.getServerBuilder().build();
+				if (serverTemp instanceof OmeroWebImageServer)
+	    			server = (OmeroWebImageServer)serverTemp;
+	    		else
+	    			Dialogs.showErrorMessage("Not an OMERO image", "Chosen image does not come from an OMERO server");
+			} catch (Exception ex) {
+				Dialogs.showErrorMessage("Error", "An error occurred while processing " + entry);
+			}
+    	}
+    	
+    	
+    	
 		mainPane = new BorderPane();
 		selectedObjects = null;
-		
-		TabPane tabPane = new TabPane();
 		
 		SplitPane browsePane = new SplitPane();
 		GridPane browseLeftPane = new GridPane();
@@ -405,19 +431,16 @@ class OmeroWebImageServerBrowser {
 		PaneTools.addGridRow(browseRightPane, 1, 0, null, description);
 		PaneTools.addGridRow(browseRightPane, 2, 0, null, actionBtnPane);
 		
-		tabPane.setTabClosingPolicy(TabClosingPolicy.UNAVAILABLE);
-		tabPane.getTabs().add(new Tab("Browse", browsePane));
-		tabPane.getTabs().add(new Tab("Client", clientPane));
 
         // Anchor the tabs and progressIndicator
         AnchorPane anchor = new AnchorPane();
-        anchor.getChildren().addAll(tabPane, progressIndicator);
+        anchor.getChildren().addAll(browsePane, progressIndicator);
         AnchorPane.setTopAnchor(progressIndicator, 3.0);
         AnchorPane.setRightAnchor(progressIndicator, 5.0);
-        AnchorPane.setTopAnchor(tabPane, 1.0);
-        AnchorPane.setRightAnchor(tabPane, 1.0);
-        AnchorPane.setLeftAnchor(tabPane, 1.0);
-        AnchorPane.setBottomAnchor(tabPane, 1.0);
+        AnchorPane.setTopAnchor(browsePane, 1.0);
+        AnchorPane.setRightAnchor(browsePane, 1.0);
+        AnchorPane.setLeftAnchor(browsePane, 1.0);
+        AnchorPane.setBottomAnchor(browsePane, 1.0);
         
         // Set HGrow and VGrow
 		GridPane.setHgrow(description,  Priority.ALWAYS);
@@ -445,17 +468,16 @@ class OmeroWebImageServerBrowser {
 		mainPane.setCenter(anchor);
 		browsePane.getItems().addAll(browseLeftPane, browseRightPane);
 		
-		refreshBtn = new Button("Refresh");
-		refreshBtn.setOnAction(e -> {
-			clientPane.setTop(getClientPane());
-		});
 		
-		GridPane buttonPane = PaneTools.createColumnGridControls(refreshBtn);
-		buttonPane.setHgap(10);
-		buttonPane.setPadding(new Insets(5, 0, 5, 0));
-		
-		clientPane.setTop(getClientPane());
-		clientPane.setBottom(buttonPane);
+		Stage dialog = new Stage();
+		dialog.sizeToScene();
+		QuPathGUI qupath = QuPathGUI.getInstance();
+		if (qupath != null)
+			dialog.initOwner(QuPathGUI.getInstance().getStage());
+		dialog.setTitle("OMERO web server");
+		dialog.setScene(new Scene(mainPane));
+		dialog.setOnCloseRequest(e -> shutdownPool());
+		dialog.showAndWait();
     }
 
     
@@ -473,141 +495,6 @@ class OmeroWebImageServerBrowser {
 			logger.warn("Could not load OMERO icons.", e.getLocalizedMessage());
 		}
 		return map;
-	}
-
-
-	private BorderPane getClientPane() {
-    	BorderPane mainPane = new BorderPane();
-		GridPane serverGrid = new GridPane();
-		serverGrid.setVgap(10.0);
-		
-		var rowIndex = 0;
-		var hostClientsMap = OmeroWebClients.getAllClients();
-		for (var hostClientEntry: hostClientsMap.entrySet()) {
-			if (hostClientEntry.getValue().isEmpty())
-				continue;
-			
-			GridPane gridPane = new GridPane();
-			GridPane infoPane = new GridPane();
-			GridPane actionPane = new GridPane();
-			
-			GridPane imageServersPane = new GridPane();
-			int imageServerRow = 0;
-			var imageList = qupath.getProject().getImageList();
-			String host = hostClientEntry.getKey();
-			Optional<OmeroWebClient> clientsWithUsername = hostClientEntry.getValue().parallelStream().filter(e -> !e.getUsername().isEmpty()).findAny();
-			Label userLabel = clientsWithUsername.isEmpty() ? new Label(host) : new Label(host + " (" + clientsWithUsername.get().getUsername() + ")");
-			PaneTools.addGridRow(infoPane, 0, 0, null, userLabel);
-			
-			for (OmeroWebClient client: hostClientEntry.getValue()) {
-				// Check if client's servers haven't been deleted from project
-				if (imageList.parallelStream().anyMatch(e -> {
-					try {
-						return e.getServerURIs().iterator().next() == client.getURI();
-					} catch (Exception ex) {
-						logger.warn(ex.getLocalizedMessage());
-					}
-					return false;
-				}))
-					imageServersPane.addRow(imageServerRow++, new Label(client.getURI().toString()));
-			}
-
-			TitledPane imageServersTitledPane = new TitledPane(imageServerRow + " server(s)", imageServersPane);
-			imageServersTitledPane.setMaxWidth(Double.MAX_VALUE);
-			imageServersTitledPane.setExpanded(false);
-			Platform.runLater(() -> {
-				try {
-					// These 2 next lines help prevent NPE
-					imageServersTitledPane.applyCss();
-					imageServersTitledPane.layout();
-					imageServersTitledPane.lookup(".title").setStyle("-fx-background-color: transparent");
-					imageServersTitledPane.lookup(".title").setEffect(null);
-					imageServersTitledPane.lookup(".content").setStyle("-fx-border-color: null");
-				} catch (Exception e) {
-					logger.error("Error setting CSS style", e.getLocalizedMessage());
-				}
-			});
-
-			PaneTools.addGridRow(infoPane, 1, 0, null, imageServersTitledPane);
-
-			// Get first client in list that requires login, or the first one if all public
-			OmeroWebClient clientToPing = clientsWithUsername.isEmpty() ? hostClientEntry.getValue().get(0) : clientsWithUsername.get();
-			boolean loggedIn = clientToPing.loggedIn();
-			
-			Node state = createStateNode(clientToPing.loggedIn());
-			Button connectionBtn = loggedIn ? new Button("Log out") : new Button("Log in");
-			Button removeBtn = new Button("Remove");
-			PaneTools.addGridRow(actionPane, 0, 0, null, state, connectionBtn, removeBtn);
-			
-			connectionBtn.setOnAction(e -> {
-				if (connectionBtn.getText().equals("Log in")) {
-					boolean success = true;
-					
-					// Check again the state, in case it wasn't refreshed in time
-					if (!clientToPing.loggedIn())
-						success = OmeroWebClients.logIn(clientToPing);
-					
-					// Change text on button if connection was successful
-					if (success) {
-						connectionBtn.setText("Log out");
-						refreshBtn.fire();
-					} else
-						Dialogs.showErrorMessage("Connect to server", "Could not connect to server. Check the log for more info.");
-				} else {
-					// Check again the state, in case it wasn't refreshed in time
-					if (clientToPing.loggedIn())
-						OmeroWebClients.logOut(clientToPing);
-
-					// Change text on button
-					connectionBtn.setText("Log in");
-					refreshBtn.fire();
-				}
-			});
-			
-			removeBtn.setOnMouseClicked(e -> {
-				if (!clientsWithUsername.isEmpty() && qupath.getViewers().stream().anyMatch(viewer -> {
-							if (viewer.getServer() == null)
-								return false;
-							return viewer.getServer().getURIs().iterator().next() == clientsWithUsername.get().getURI();
-						})) {
-					Dialogs.showMessageDialog("Remove server", "You need to close the image in the viewer first!");
-					return;
-				}
-				if (!clientsWithUsername.isEmpty() && clientToPing.loggedIn())
-					OmeroWebClients.logOut(clientToPing);
-				OmeroWebClients.removeClient(clientToPing);
-				refreshBtn.fire();
-			});
-			
-			connectionBtn.setDisable(clientsWithUsername.isEmpty());
-			if (!clientsWithUsername.isEmpty())
-				removeBtn.disableProperty().bind(connectionBtn.textProperty().isEqualTo("Log out"));
-			else
-				removeBtn.setDisable(false);
-			
-			PaneTools.addGridRow(gridPane, 0, 0, null, infoPane);
-			PaneTools.addGridRow(gridPane, 0, 1, null, actionPane);
-			
-			GridPane.setHgrow(gridPane, Priority.ALWAYS);
-			GridPane.setHgrow(actionPane, Priority.ALWAYS);
-			actionPane.setAlignment(Pos.CENTER_RIGHT);
-
-			gridPane.setStyle("-fx-border-color: black;");
-			serverGrid.add(gridPane, 0, rowIndex++);
-		}
-		
-		if (serverGrid.getChildren().isEmpty())
-			serverGrid.add(new Label("No OMERO server"), 0, 0);
-			
-		serverGrid.setMinWidth(250);
-		mainPane.setTop(serverGrid);
-		return mainPane;
-    }
-
-
-	private Node createStateNode(boolean loggedIn) {
-		var state = loggedIn ? IconFactory.PathIcons.ACTIVE_SERVER : IconFactory.PathIcons.INACTIVE_SERVER;
-		return IconFactory.createNode(QuPathGUI.TOOLBAR_ICON_SIZE, QuPathGUI.TOOLBAR_ICON_SIZE, state);
 	}
 
 
@@ -1448,10 +1335,4 @@ class OmeroWebImageServerBrowser {
 	public void shutdownPool() {
 		executorTable.shutdownNow();
 	}
-	
-	public BorderPane getPane() {
-		return this.mainPane;
-	}
-
-
 }
