+++ qupath-extension-omero/src/main/resources/META-INF/services/qupath.lib.gui.extensions.QuPathExtension
@@ -1,0 +1,1 @@
+qupath.lib.images.servers.omero.OmeroExtension
